requirements.txt
streamlit>=1.30.0
langchain>=0.1.14
langchain-community>=0.0.11
openai>=1.14.0
python-dotenv>=1.0.0
Pillow>=10.2.0